+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Adding `xan slice -E/--end-byte`.
 * Allowing `-n/--no-headers` and `-d/--delimiter` flags to appear beforecolorize_main_help subcommands.
 * Adding backtick quoted strings to moonblade.
+* Adding moonblade `printf` function.
 
 *Fixes*